+++ src/akismet/_async_client.py
@@ -280,7 +280,7 @@ class AsyncClient:
         version: str,
         endpoint: str,
         user_ip: str,
-        **kwargs: Unpack[_common.AkismetArguments],
+        **kwargs: Unpack[akismet.AkismetArguments],
     ) -> httpx.Response:
         """
         Make a POST request to the Akismet API and return the response.
@@ -315,7 +315,7 @@ class AsyncClient:
         )
 
     async def _submit(
-        self, endpoint: str, user_ip: str, **kwargs: Unpack[_common.AkismetArguments]
+        self, endpoint: str, user_ip: str, **kwargs: Unpack[akismet.AkismetArguments]
     ) -> bool:
         """
         Submit ham or spam to the Akismet API.
@@ -340,7 +340,7 @@ class AsyncClient:
     # ----------------------------------------------------------------------------
 
     async def comment_check(
-        self, user_ip: str, **kwargs: Unpack[_common.AkismetArguments]
+        self, user_ip: str, **kwargs: Unpack[akismet.AkismetArguments]
     ) -> akismet.CheckResponse:
         """
         Check a piece of user-submitted content to determine whether it is spam.
@@ -391,7 +391,7 @@ class AsyncClient:
         )
 
     async def submit_ham(
-        self, user_ip: str, **kwargs: Unpack[_common.AkismetArguments]
+        self, user_ip: str, **kwargs: Unpack[akismet.AkismetArguments]
     ) -> bool:
         """
         Inform Akismet that a piece of user-submitted comment is not spam.
@@ -433,7 +433,7 @@ class AsyncClient:
         return await self._submit(_common._SUBMIT_HAM, user_ip, **kwargs)
 
     async def submit_spam(
-        self, user_ip: str, **kwargs: Unpack[_common.AkismetArguments]
+        self, user_ip: str, **kwargs: Unpack[akismet.AkismetArguments]
     ) -> bool:
         """
         Inform Akismet that a piece of user-submitted comment is spam.